Refactor expression converters so we don't repeat ourselves or worse, repeat ourselves
