+++ orquestaconvert/expressions/base.py
@@ -1,5 +1,15 @@
 import abc
+import re
 import six
+
+# task('xxx').result -> result()
+#    if 'xxx' != current task name, error
+TASK_RESULT_REGEX = r"(task\([\"\']?\w+[\"\']?\).result)"
+TASK_RESULT_PATTERN = re.compile(TASK_RESULT_REGEX)
+
+# - st2kv. -> st2kv('xxx')
+ST2KV_REGEX = r"(st2kv\.([\w\.]+))"
+ST2KV_PATTERN = re.compile(ST2KV_REGEX)
 
 
 @six.add_metaclass(abc.ABCMeta)
@@ -75,5 +85,14 @@ class BaseExpressionConverter(AbstractBaseExpressionConverter):
         return "result()"
 
     @classmethod
+    def convert_task_result(cls, expr):
+        # TODO error if task name is not the same task in this context
+        return TASK_RESULT_PATTERN.sub(cls._replace_task_result, expr)
+
+    @classmethod
     def _replace_st2kv(cls, match):
         return "st2kv('" + match.group(2) + "')"
+
+    @classmethod
+    def convert_st2kv(cls, expr):
+        return ST2KV_PATTERN.sub(cls._replace_st2kv, expr)
